camxes-bot and cipra-bot respond to their nickname
Up until now, camxes-bot was responding to the nick “camxes”, even if it
was named otherwise in the config. So did cipra-bot, which responded
only to “cipra”. This resulted in incomprehensions for users if the bot
had another nickname.

For instance, with a bot named “Barfoo”, the following commands would be
ignored:

baz> barfoo: bangu
baz> coi la .barfoo.

But these would be responded to:

baz> camxes: bangu
barfoo> (bangu VAU)
baz> coi la .camxes.
barfoo> coi

Now, it responds to it's name:

baz> barfoo: bangu
barfoo> (bangu VAU)
baz> coi la .barfoo.
barfoo> coi

There is still a bug, when the nickname is already used on the IRC
network on which camxes-bot is connected. The bot is then named
differently (say, “barfoo1” instead of “barfoo”), but still respond only
to “barfoo”. This is however less surprising than the former behavior,
and less simple to correct.