Prevent a crash when ~/.livla/notci.txt is missing
When launching the bot, it crashed if “~/.livla/notci.txt” did not
exist. When “~/.livla/” did not exist, this problem repeated each
launch.

When “~/.livla/” did exist, it succesfully created the “notci.txt” file,
then crashed, due to asynchronous business.

The script now synchronously create “~/.livla/notci.txt”, and keep going
after that.

Note that we abandonned “fs.exists” for “fs.appendSync”.

From the node.js documentation:

  fs.exists() is an anachronism and exists only for historical reasons.
  There should almost never be a reason to use it in your own code.

  (http://nodejs.org/api/fs.html#fs_fs_exists_path_callback)

“fs.appendSync” will create the file if necessary, and will only “touch”
it, without modifying it, when it is already there.

“fs.mkdirSync” will complain if the directory already exists. Following
the aforementioned documentation, we treat this case with error handling
(saying “The directory existing is not an error”), rather than
preventively testing for it's existence.

On real errors (like, no write permissions – even if the file exists),
node will throw, hopefully with a sensible error message.