Added npm package.json for module management
Some of the modules that come with this repository are compiled for a 64
bit system (actually, “libxmljs” was the culprit). Which was not good on
my 32 bit system.

An easy solution to fix this is to remove the node_modules altogether
and regenerate it with a simple “npm install”. That is, when we have a
correct “package.json”.

So, I created one (well, two, one for the main repository, and one for
“ircbot”, which has it's own “node_modules” directory).

I also moved “node-supervisor” so that it respects the installation
scheme, and that we can really remove and reinstall the node_modules
directory, without having to manage this weird “node-supervisor”
exception.

This required a small change in the command to launch (by default,
binaries are in “node_modules/.bin”, not “node_modules/bin”).

While I was at it, since we still provide the packages in the
repositories, I added a note in the same readme so that people on a 32
bit system know what to do.