Corrected documentation about livlytcan
Well, not connecting to livlytcan doesn't provoque a crash.

But it seems that the replier dos not respond anymore if we do so.

So, I corrected the documentation accordingly.

The bots don't seem to speak on livlytcan, but I guess they need that
channel to be able to contact themselves in a private room?

I'll have to dig more. But I don't want to :-O.