Moved “require("fs")” to the top of livlabot.
The line `var fs = require("fs"), path = require("path-extra");`
appeared again and again throughout the file. Never, in the file, are
`fs` and `path` defined to be anything else than these modules.

We can load the modules once and for all, and use them in the script
without ensuring that they are indeed these modules.