Created a function to create the config directory.
I created a function to create the config directory. Or any directory,
for that matter. It essentially does what was done when loading
“~/.livla/notci.txt”, directory wise: try to create it, and if it fails,
check that the directory already exist.

One difference is that the previous code was not failing if “~/.livla”
existed but was not a directory. This code will.