Created a function to read configuration files.
The `readConfig` function is here to read any file situated in
“~/.livla”. If the file does not exist, it returns an empty string.
`readConfig` ensure that “~/.livla” exists.

`notci` is now loaded with `readConfig`. One difference with the code
before this commit is that `readConfig` does not create the missing
files. So, unless we have something to put in it, “notci.txt” is not
created when we launch the script. However, this file, once loaded, is
only written in, never read. So, this should not cause any problem.