Updated LICENSE with author names
Masato only wrote the “camxes.js.peg” and “camxes.js” part of the
project.

The difference beetween the license of his work, and the license of the
project, is now clearly stated.

The camxes.js.peg is updated from the Masato repository, where he states
the license of his file in the file itself (which is good practice).